feat(response): segregate ResponseInterface and un-sibling IBS from CNR
Introduce CNIC\AbstractResponse and CNIC\AbstractRecord as brand-neutral
bases. CNR\Response and IBS\Response (and their Records) now extend them as
siblings instead of IBS extending CNR, removing the false is-a relationship.

The CNR-only capabilities (getQueuetime/getRuntime/isTmpError/isPending/
getListHash) move off the core ResponseInterface into a new
CNIC\ExtendedResponseInterface implemented only by CNR\Response, so IBS/Moniker
responses no longer implement-and-throw those methods — they are simply absent.
This retires the refused-bequest smell tracked in RSRMID-2904.

BREAKING CHANGE: getQueuetime(), getRuntime(), isTmpError(), isPending() and
getListHash() are no longer part of CNIC\ResponseInterface. Consumers holding
the core interface (e.g. the AbstractClient::request() return type) must narrow
via `instanceof CNIC\ExtendedResponseInterface` before calling them; consumers
holding the concrete CNR\Client/CNR\Response are unaffected. IBS\Response no
longer extends CNR\Response and IBS\Record no longer extends CNR\Record.
