+++ src/CNR/Record.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 
 namespace CNIC\CNR;
 
+use CNIC\AbstractRecord;
 use CNIC\RecordInterface;
 
 /**
@@ -17,53 +18,6 @@ use CNIC\RecordInterface;
  * @psalm-api
  * @package CNIC\CNR
  */
-class Record implements RecordInterface
+class Record extends AbstractRecord implements RecordInterface
 {
-    /**
-     * Constructor
-     * e.g.
-     * <code>
-     * $data = [
-     *   "DOMAIN" => "mydomain.com",
-     *   "USER"   => "test.user",
-     *   // ... further column data ...
-     * ];
-     * </code>
-     * @param array<string,mixed> $data row data
-     */
-    public function __construct(private readonly array $data)
-    {
-    }
-
-    /**
-     * get row data
-     * @return array<string,mixed>
-     */
-    #[\Override]
-    public function getData(): array
-    {
-        return $this->data;
-    }
-
-    /**
-     * get row data for given column
-     * @param string $key column name
-     */
-    #[\Override]
-    public function getDataByKey(string $key): mixed
-    {
-        if ($this->hasData($key)) {
-            return $this->data[$key];
-        }
-        return null;
-    }
-
-    /**
-     * check if record has data for given column
-     * @param string $key column name
-     */
-    private function hasData(string $key): bool
-    {
-        return array_key_exists($key, $this->data);
-    }
 }
