+++ src/ResponseInterface.php
@@ -50,22 +50,10 @@ interface ResponseInterface
     public function getPlain(): string;
 
     /**
-     * Get Queuetime of API response
-     * @return float Queuetime of API response
-     */
-    public function getQueuetime(): float;
-
-    /**
      * Get API response as Hash
      * @return array<string, mixed> API response hash
      */
     public function getHash(): array;
-
-    /**
-     * Get Runtime of API response
-     * @return float Runtime of API response
-     */
-    public function getRuntime(): float;
 
     /**
      * Check if current API response represents an error case
@@ -80,19 +68,6 @@ interface ResponseInterface
      * @return bool boolean result
      */
     public function isSuccess(): bool;
-
-    /**
-     * Check if current API response represents a temporary error case
-     * API response code is an 4xx code
-     * @return bool result
-     */
-    public function isTmpError(): bool;
-
-    /**
-     * Check if current operation is returned as pending
-     * @return bool result
-     */
-    public function isPending(): bool;
 
     /**
      * Add a column to the column list
@@ -175,12 +150,6 @@ interface ResponseInterface
      * @return int|null record index or null for a non-list response
      */
     public function getLastRecordIndex(): ?int;
-
-    /**
-     * Get Response as List Hash including useful meta data for tables
-     * @return array<mixed> hash including list meta data and array of rows in hash notation
-     */
-    public function getListHash(): array;
 
     /**
      * Get next record in record list
